Chore: CI 허스키, 패키지 설정 및 TS 프로젝트로 변경
