+++ .eslintrc
@@ -1,0 +1,11 @@
+{
+  "extends": ["react-app", "eslint:recommended"],
+  "rules": {
+    "no-var": "error",
+    "no-multiple-empty-lines": "error",
+    "no-console": ["warn", { "allow": ["warn", "error", "info"] }],
+    "eqeqeq": "error",
+    "no-unused-vars": "warn",
+    "no-undef": "warn"
+  }
+}
