+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,16 @@
+## 🔎 What is this PR?
+
+- 해당 PR에 대한 설명을 적어주세요.
+-
+-
+
+## 🔑 Key Changes
+
+- 핵심 변경 내용이나 코드에 대해 설명해주세요.
+-
+-
+
+## 🙋🏻‍♀️ To Reviewers
+
+- 코드 리뷰시 집중할 부분에 대해 설명해주세요.
+-
