+++ .prettierrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  singleQuote: true,
+  semi: true,
+  useTabs: false,
+  tabWidth: 2,
+  trailingComma: 'all',
+  printWidth: 80,
+};
