+++ src/App.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import CommentListContainer from './containers/CommentListContainer';
+import FormContainer from './containers/FormContainer';
+import PageListContainer from './containers/PageListContainer';
+
+function App() {
+  return (
+    <div>
+      <CommentListContainer />
+      <PageListContainer />
+      <FormContainer />
+    </div>
+  );
+}
+
+export default App;
