+++ webpack.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  resolve: {
+    extensions: ['.js', '.jsx', '.ts', '.tsx'],
+  },
+};
